Check for login method instead of active session id.
Previously would add the session id if set.  This would cause a problem in
cases where the class was not reinstantiated on each login.

Problem was found in the Ruby version since that is used in a long running
process.  Ported fix to PHP version for consistency.

Ref: PHP SOAP Client library

Committed by: Paul Barrett